+++ checkout/src/main/java/com/mercadopago/sdk/android/checkout/domain/extensions/FieldStateExtensions.kt
@@ -1,0 +1,10 @@
+package com.mercadopago.sdk.android.checkout.domain.extensions
+
+import com.mercadopago.sdk.android.checkout.presentation.state.CardNumberState
+import com.mercadopago.sdk.android.checkout.presentation.state.IdentificationTypeState
+
+internal fun CardNumberState.isComplete() = length == maxLength
+
+internal fun IdentificationTypeState.isComplete(
+    length: Int,
+) = selected?.maxLength == length || selected?.minLength == length
