+++ checkout/src/main/java/com/mercadopago/sdk/android/checkout/domain/usecase/GetCardDataByBinUseCase.kt
@@ -5,13 +5,13 @@ import com.mercadopago.sdk.android.checkout.core.model.CardType
 import com.mercadopago.sdk.android.checkout.core.model.PaymentMethod
 import com.mercadopago.sdk.android.checkout.domain.exception.ErrorCode
 import com.mercadopago.sdk.android.checkout.domain.exception.ErrorLocalized
+import com.mercadopago.sdk.android.checkout.domain.extensions.extractCardFilters
 import com.mercadopago.sdk.android.checkout.domain.extensions.flatMap
 import com.mercadopago.sdk.android.checkout.domain.extensions.fold
-import com.mercadopago.sdk.android.checkout.domain.mapper.extractCardFilters
-import com.mercadopago.sdk.android.checkout.domain.mapper.hasIssuers
-import com.mercadopago.sdk.android.checkout.domain.mapper.matchesCardBrand
-import com.mercadopago.sdk.android.checkout.domain.mapper.matchesCardType
-import com.mercadopago.sdk.android.checkout.domain.mapper.toSecurityCode
+import com.mercadopago.sdk.android.checkout.domain.extensions.hasIssuers
+import com.mercadopago.sdk.android.checkout.domain.extensions.matchesCardBrand
+import com.mercadopago.sdk.android.checkout.domain.extensions.matchesCardType
+import com.mercadopago.sdk.android.checkout.domain.extensions.toSecurityCode
 import com.mercadopago.sdk.android.checkout.domain.model.CardData
 import com.mercadopago.sdk.android.checkout.domain.model.MercadoPagoCheckoutError
 import com.mercadopago.sdk.android.coremethods.domain.utils.Result
